creation of the MetaData.php file, file extended to the Products and Users file
